linear optimization method for interventional CIMs
QIG.py is used to generate QIG objects from either observational or a  combination of observational and interventional data. The constructed QIG object includes the data vector for integer linear programming methods for causal discovery, an estimated tree skeleton, as well as the defining hyperplanes of the associated CIM polytope. Included also are simulations for observational data and a real data example using the Sachs et al. data set.